Simple higher level garbage collector added.
This is a very simple interface that runs a threaded `Collector` that
calls a `collect` method on all items that are being tracked. If the
method returns True the `Collector` assumes everything was cleaned and
will remove the object from tracking.

Each item is a subclass of `Garbage`.

This instance is ment to keep track of incorrectly closed threads,
subprocesses or other pieces that don't always clean properly.