Moved IRC related files to their own package named 'bot'
This is more cleaning up of the root directory of this project. The
IRC bot functionality is now all contained in the 'bot' subpackage.

This also moves the IRC communication from multiprocessing.managers
to the `jsonrpclib` already used by the queue. This means announcers
need to call functions exported in the `bot` package. They're
available from the `__init__.py`.

Due to the API change, we've also updated the API calls in `requests_`
and `manager.status`.

Starting the IRC bot should still be done with the root directory of
this project as CWD due to configuration compatibility. Run the
following command from there `python -m bot.runner` until we are
able to migrate to a sane configuration method.