A whole bunch of changes to listeners handling
Bottom line is, streamstatus now fetches a listener
aggregate from the load balancer, if the load balancer
is responding (which it always should be).

I don't really know what it does with regard to player
types for player statistics, but that's fucked anyway
since we don't forward the correct IP address to ice.